third_party/blink/web_tests/platform/mac/external/wpt/mathml/presentation-markup/operators/stretchy-largeop-with-default-font-2-expected.txt

This is a testharness.js-based test.
[FAIL] Test that non-stretchy largeops are bigger in display mode.
  assert_greater_than: size of largeop '⨀' expected a number greater than 12.359375 but got 12.359375
Harness: the test ran to completion.
